Rails 5 upgrade

+++ db/migrate/20200331063308_remove_extended_product_from_shopify_cred.rb
@@ -1,6 +1,6 @@
-class RemoveExtendedProductFromShopifyCred < ActiveRecord::Migration
+class RemoveExtendedProductFromShopifyCred < ActiveRecord::Migration[5.1]
   def up
-    remove_column :shipping_easy_credentials, :includes_product
+    remove_column :shipping_easy_credentials, :includes_product if column_exists? :shipping_easy_credentials, :includes_product
   end
 
   def down
